\name{dataPop_1999_2020}
\docType{data}
\alias{dataPop_1999_2020}
\title{Population data (1999-2020)}
\description{
  This data set contains mortality rate for various demographic groups by age. all
  column has overall mortality rate for all groups. Race groups are indicated by the
  following notation: w - white, b - black, ai - american indian and as - asian.
  Female and Males are indicated by an additional suffix _f and _m.
}
\usage{dataPop_1999_2020}
\format{A dataframe with 16 columns and 151 rows.}
\source{CDC Wonder Database. Data Colleceted from 1999 - 2020}
\references{
  https://wonder.cdc.gov/
}
\keyword{datasets}