\name{dataPop_2018_2021}
\docType{data}
\alias{dataPop_2018_2021}
\title{Population data (2018_2021)}
\description{
  This data set contains mortality rate for various demographic groups by age. all
  column has overall mortality rate for all groups. Race groups are indicated by the
  following notation: w - white, b - black, ai - american indian and as - asian, nh - native hawaiian.
  Female and Males are indicated by an additional suffix _f and _m.
}
\usage{dataPop_2018_2021}
\format{A dataframe with 19 columns and 151 rows.}
\source{CDC Wonder Database. Data Colleceted from 2018_2021}
\references{
  https://wonder.cdc.gov/
}
\keyword{datasets}