\name{dataSurv}
\docType{data}
\alias{dataSurv}
\title{Survival data}
\description{
  This data set is obtained from Finkelstein et al. paper that contains the following
  five columns: age, time, event status, sex and race.
}
\usage{dataSurv}
\format{A dataframe with 5 columns and 33 rows.}
\source{Finkelstein et al. (2003)}
\references{
  Finkelstein, D. M., Muzikansky, A., & Schoenfeld, D. A. (2003). Comparing survival of a sample to that of a standard population. Journal of the National Cancer Institute, 95(19), 1434-1439.
}
\keyword{datasets}