\name{dataSurv_small}
\docType{data}
\alias{dataSurv_small}
\title{Survival data}
\description{
  This data set is subset of data obtained from Finkelstein et al. paper that contains the following
  five columns: age, time, event status, sex and race. In order to apply the 
  exact test 12 patients were randomly selected out of 33 patients.
}
\usage{dataSurv}
\format{A dataframe with 5 columns and 12 rows.}
\source{Finkelstein et al. (2003)}
\references{
  Finkelstein, D. M., Muzikansky, A., & Schoenfeld, D. A. (2003). Comparing survival of a sample to that of a standard population. Journal of the National Cancer Institute, 95(19), 1434-1439.
}
\keyword{datasets}