% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneSampleLogRankTest.R
\name{oneSampleLogRankTest}
\alias{oneSampleLogRankTest}
\title{Calculate One-Sample Log-Rank Test}
\usage{
oneSampleLogRankTest(dataSurv, dataPop, type = c("exact", "approximate"))
}
\arguments{
\item{dataSurv}{Survival data}

\item{dataPop}{Population data}

\item{type}{Type of test}
}
\value{
p-value for one-sample log-rank test
}
\description{
Calculate One-Sample Log-Rank Test
}
\examples{
# load data
data(dataSurv_small)
data(dataPop_2018_2021)

# Since the dataset is small run an exact test
oneSampleLogRankTest(dataSurv_small, dataPop_2018_2021, type = "exact")

}
