% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneSampleLogRankTest.R
\name{plotKM}
\alias{plotKM}
\title{Plot Kaplan-Meier Curve against Population}
\usage{
plotKM(dataSurv, dataPop, type = c("exact", "approximate"))
}
\arguments{
\item{dataSurv}{Survival data}

\item{dataPop}{Population data}

\item{type}{Type of test to conduct in order to display p-value}
}
\value{
ggplot object
}
\description{
Plot Kaplan-Meier Curve against Population
}
\examples{
# load data
data(dataSurv_small)
data(dataPop_2018_2021)

plotKM(dataSurv_small, dataPop_2018_2021, type = "exact")

}
