\name{simulated_clinical_data}
\docType{data}
\alias{simulated_clinical_data}
\title{Simultaed Clinical data}
\description{
  This data is simulated data from clinical trial data that contains five columns:
  race, sex, age, event status and time in years.
}
\usage{simulated_clinical_data}
\format{A dataframe with 5 columns and 500 rows.}
\source{Simulated}
\references{None}
\keyword{datasets}